perf(desktop): lazy-load ApprovalModal and ContextPanel
Extract ContextPanel's pure utility functions into contextPanelUtils.ts
so initial-path consumers (ContextWindowRing → Composer) can import them
without pulling the full React component into the initial bundle.

ApprovalModal and ContextPanel are both conditionally rendered, making
them safe lazy-load candidates following the existing RemotePanel pattern.

Initial JS gzip: 472.7 → 464.1 KiB (−8.6 KiB).

将 ContextPanel 的纯工具函数提取到 contextPanelUtils.ts，使初始路径消费者
（ContextWindowRing → Composer）可以导入它们而不会将完整的 React 组件拉入
初始包。ApprovalModal 和 ContextPanel 均为条件渲染，遵循已有的 RemotePanel
模式安全地 lazy 化。初始 JS gzip：472.7 → 464.1 KiB（−8.6 KiB）。
